+++ mkdocs/commands/serve.py
@@ -1,14 +1,33 @@
-from __future__ import unicode_literals
-
 import logging
 import shutil
 import tempfile
+import sys
 
 from os.path import isfile, join
 from mkdocs.commands.build import build
 from mkdocs.config import load_config
 
 log = logging.getLogger(__name__)
+
+
+def _init_asyncio_patch():
+    """
+    Select compatible event loop for Tornado 5+.
+
+    As of Python 3.8, the default event loop on Windows is `proactor`,
+    however Tornado requires the old default "selector" event loop.
+    As Tornado has decided to leave this to users to set, MkDocs needs
+    to set it. See https://github.com/tornadoweb/tornado/issues/2608.
+    """
+    if sys.platform.startswith("win") and sys.version_info >= (3, 8):
+        import asyncio
+        try:
+            from asyncio import WindowsSelectorEventLoopPolicy
+        except ImportError:
+            pass  # Can't assign a policy which doesn't exist.
+        else:
+            if not isinstance(asyncio.get_event_loop_policy(), WindowsSelectorEventLoopPolicy):
+                asyncio.set_event_loop_policy(WindowsSelectorEventLoopPolicy())
 
 
 def _get_handler(site_dir, StaticFileHandler):
@@ -24,7 +43,7 @@ def _get_handler(site_dir, StaticFileHandler):
                 if isfile(join(site_dir, error_page)):
                     self.write(Loader(site_dir).load(error_page).generate())
                 else:
-                    super(WebHandler, self).write_error(status_code, **kwargs)
+                    super().write_error(status_code, **kwargs)
 
     return WebHandler
 
@@ -33,13 +52,14 @@ def _livereload(host, port, config, builder, site_dir):
 
     # We are importing here for anyone that has issues with livereload. Even if
     # this fails, the --no-livereload alternative should still work.
+    _init_asyncio_patch()
     from livereload import Server
     import livereload.handlers
 
     class LiveReloadServer(Server):
 
         def get_web_handlers(self, script):
-            handlers = super(LiveReloadServer, self).get_web_handlers(script)
+            handlers = super().get_web_handlers(script)
             # replace livereload handler
             return [(handlers[0][0], _get_handler(site_dir, livereload.handlers.StaticFileHandler), handlers[0][2],)]
 
@@ -62,6 +82,7 @@ def _static_server(host, port, site_dir):
 
     # Importing here to seperate the code paths from the --livereload
     # alternative.
+    _init_asyncio_patch()
     from tornado import ioloop
     from tornado import web
 
@@ -82,7 +103,7 @@ def _static_server(host, port, site_dir):
 
 
 def serve(config_file=None, dev_addr=None, strict=None, theme=None,
-          theme_dir=None, livereload='livereload'):
+          theme_dir=None, livereload='livereload', **kwargs):
     """
     Start the MkDocs development server
 
@@ -104,10 +125,11 @@ def serve(config_file=None, dev_addr=None, strict=None, theme=None,
             strict=strict,
             theme=theme,
             theme_dir=theme_dir,
-            site_dir=site_dir
+            site_dir=site_dir,
+            **kwargs
         )
         # Override a few config settings after validation
-        config['site_url'] = 'http://{0}/'.format(config['dev_addr'])
+        config['site_url'] = 'http://{}/'.format(config['dev_addr'])
 
         live_server = livereload in ['dirty', 'livereload']
         dirty = livereload == 'dirty'