+++ mkdocs/tests/config/base_tests.py
@@ -1,15 +1,9 @@
-from __future__ import unicode_literals
 import os
 import tempfile
 import unittest
+from tempfile import TemporaryDirectory
 
-try:
-    # py>=3.2
-    from tempfile import TemporaryDirectory
-except ImportError:
-    from backports.tempfile import TemporaryDirectory
-
-from mkdocs import exceptions, utils
+from mkdocs import exceptions
 from mkdocs.config import base, defaults
 from mkdocs.config.config_options import BaseConfigOption
 
@@ -274,6 +268,6 @@ class ConfigBaseTests(unittest.TestCase):
             self.assertEqual(cfg['site_name'], 'MkDocs Test')
             self.assertEqual(cfg['docs_dir'], docs_dir)
             self.assertEqual(cfg.config_file_path, config_fname)
-            self.assertIsInstance(cfg.config_file_path, utils.text_type)
+            self.assertIsInstance(cfg.config_file_path, str)
         finally:
             config_dir.cleanup()