+++ mkdocs/tests/gh_deploy_tests.py
@@ -1,14 +1,26 @@
-from __future__ import unicode_literals
-
 import unittest
-import mock
+from unittest import mock
 
-from mkdocs.tests.base import load_config, LogTestCase
+from mkdocs.tests.base import load_config
 from mkdocs.commands import gh_deploy
 from mkdocs import __version__
 
 
 class TestGitHubDeploy(unittest.TestCase):
+
+    def assert_mock_called_once(self, mock):
+        """assert that the mock was called only once.
+
+        The `mock.assert_called_once()` method was added in PY36.
+        TODO: Remove this when PY35 support is dropped.
+        """
+        try:
+            mock.assert_called_once()
+        except AttributeError:
+            if not mock.call_count == 1:
+                msg = ("Expected '%s' to have been called once. Called %s times." %
+                       (mock._mock_name or 'mock', self.call_count))
+                raise AssertionError(msg)
 
     @mock.patch('subprocess.Popen')
     def test_is_cwd_git_repo(self, mock_popeno):
@@ -29,7 +41,7 @@ class TestGitHubDeploy(unittest.TestCase):
 
         mock_popeno().communicate.return_value = (b'6d98394\n', b'')
 
-        self.assertEqual(gh_deploy._get_current_sha('.'), u'6d98394')
+        self.assertEqual(gh_deploy._get_current_sha('.'), '6d98394')
 
     @mock.patch('subprocess.Popen')
     def test_get_remote_url_ssh(self, mock_popeno):
@@ -39,7 +51,7 @@ class TestGitHubDeploy(unittest.TestCase):
             b''
         )
 
-        expected = (u'git@', u'mkdocs/mkdocs.git')
+        expected = ('git@', 'mkdocs/mkdocs.git')
         self.assertEqual(expected, gh_deploy._get_remote_url('origin'))
 
     @mock.patch('subprocess.Popen')
@@ -50,7 +62,7 @@ class TestGitHubDeploy(unittest.TestCase):
             b''
         )
 
-        expected = (u'https://', u'mkdocs/mkdocs.git')
+        expected = ('https://', 'mkdocs/mkdocs.git')
         self.assertEqual(expected, gh_deploy._get_remote_url('origin'))
 
     @mock.patch('subprocess.Popen')
@@ -93,7 +105,7 @@ class TestGitHubDeploy(unittest.TestCase):
     @mock.patch('mkdocs.commands.gh_deploy._is_cwd_git_repo', return_value=True)
     @mock.patch('mkdocs.commands.gh_deploy._get_current_sha', return_value='shashas')
     @mock.patch('mkdocs.commands.gh_deploy._get_remote_url', return_value=(
-        u'git@', u'mkdocs/mkdocs.git'))
+        'git@', 'mkdocs/mkdocs.git'))
     @mock.patch('mkdocs.commands.gh_deploy._check_version')
     @mock.patch('mkdocs.commands.gh_deploy.ghp_import.ghp_import', return_value=(True, ''))
     def test_deploy_hostname(self, mock_import, check_version, get_remote, get_sha, is_repo):
@@ -114,7 +126,7 @@ class TestGitHubDeploy(unittest.TestCase):
             remote_branch='test',
         )
         gh_deploy.gh_deploy(config)
-        check_version.assert_called_once()
+        self.assert_mock_called_once(check_version)
 
     @mock.patch('mkdocs.commands.gh_deploy._is_cwd_git_repo', return_value=True)
     @mock.patch('mkdocs.commands.gh_deploy._get_current_sha', return_value='shashas')
@@ -147,7 +159,7 @@ class TestGitHubDeploy(unittest.TestCase):
                                                error_string)
 
 
-class TestGitHubDeployLogs(LogTestCase):
+class TestGitHubDeployLogs(unittest.TestCase):
 
     @mock.patch('subprocess.Popen')
     def test_mkdocs_newer(self, mock_popeno):