+++ mkdocs/tests/plugin_tests.py
@@ -1,20 +1,17 @@
 #!/usr/bin/env python
-# coding: utf-8
 
-from __future__ import unicode_literals
 
 import unittest
-import mock
+from unittest import mock
 import os
 
 from mkdocs import plugins
-from mkdocs import utils
 from mkdocs import config
 
 
 class DummyPlugin(plugins.BasePlugin):
     config_scheme = (
-        ('foo', config.config_options.Type(utils.string_types, default='default foo')),
+        ('foo', config.config_options.Type(str, default='default foo')),
         ('bar', config.config_options.Type(int, default=0)),
         ('dir', config.config_options.Dir(exists=False)),
     )