+++ mkdocs/tests/structure/page_tests.py
@@ -1,20 +1,12 @@
-from __future__ import unicode_literals
-
 import unittest
 import os
 import sys
-import mock
-import io
-
-try:
-    # py>=3.2
-    from tempfile import TemporaryDirectory
-except ImportError:
-    from backports.tempfile import TemporaryDirectory
+from unittest import mock
+from tempfile import TemporaryDirectory
 
 from mkdocs.structure.pages import Page
 from mkdocs.structure.files import File, Files
-from mkdocs.tests.base import load_config, dedent, LogTestCase
+from mkdocs.tests.base import load_config, dedent
 
 
 class PageTests(unittest.TestCase):
@@ -66,6 +58,54 @@ class PageTests(unittest.TestCase):
         self.assertEqual(pg.meta, {})
         self.assertEqual(pg.next_page, None)
         self.assertEqual(pg.parent, 'foo')
+        self.assertEqual(pg.previous_page, None)
+        self.assertEqual(pg.title, 'Foo')
+        self.assertEqual(pg.toc, [])
+
+    def test_nested_index_page_no_parent(self):
+        cfg = load_config(docs_dir=self.DOCS_DIR)
+        fl = File('sub1/index.md', cfg['docs_dir'], cfg['site_dir'], cfg['use_directory_urls'])
+        pg = Page('Foo', fl, cfg)
+        pg.parent = None  # non-homepage at nav root level; see #1919.
+        self.assertEqual(pg.url, 'sub1/')
+        self.assertEqual(pg.abs_url, None)
+        self.assertEqual(pg.canonical_url, None)
+        self.assertEqual(pg.edit_url, None)
+        self.assertEqual(pg.file, fl)
+        self.assertEqual(pg.content, None)
+        self.assertFalse(pg.is_homepage)
+        self.assertTrue(pg.is_index)
+        self.assertTrue(pg.is_page)
+        self.assertFalse(pg.is_section)
+        self.assertTrue(pg.is_top_level)
+        self.assertEqual(pg.markdown, None)
+        self.assertEqual(pg.meta, {})
+        self.assertEqual(pg.next_page, None)
+        self.assertEqual(pg.parent, None)
+        self.assertEqual(pg.previous_page, None)
+        self.assertEqual(pg.title, 'Foo')
+        self.assertEqual(pg.toc, [])
+
+    def test_nested_index_page_no_parent_no_directory_urls(self):
+        cfg = load_config(docs_dir=self.DOCS_DIR, use_directory_urls=False)
+        fl = File('sub1/index.md', cfg['docs_dir'], cfg['site_dir'], cfg['use_directory_urls'])
+        pg = Page('Foo', fl, cfg)
+        pg.parent = None  # non-homepage at nav root level; see #1919.
+        self.assertEqual(pg.url, 'sub1/index.html')
+        self.assertEqual(pg.abs_url, None)
+        self.assertEqual(pg.canonical_url, None)
+        self.assertEqual(pg.edit_url, None)
+        self.assertEqual(pg.file, fl)
+        self.assertEqual(pg.content, None)
+        self.assertFalse(pg.is_homepage)
+        self.assertTrue(pg.is_index)
+        self.assertTrue(pg.is_page)
+        self.assertFalse(pg.is_section)
+        self.assertTrue(pg.is_top_level)
+        self.assertEqual(pg.markdown, None)
+        self.assertEqual(pg.meta, {})
+        self.assertEqual(pg.next_page, None)
+        self.assertEqual(pg.parent, None)
         self.assertEqual(pg.previous_page, None)
         self.assertEqual(pg.title, 'Foo')
         self.assertEqual(pg.toc, [])
@@ -378,7 +418,7 @@ class PageTests(unittest.TestCase):
             fl = File('index.md', cfg['docs_dir'], cfg['site_dir'], cfg['use_directory_urls'])
             pg = Page(None, fl, cfg)
             # Create an UTF-8 Encoded file with BOM (as Micorsoft editors do). See #1186
-            with io.open(fl.abs_src_path, 'w', encoding='utf-8-sig') as f:
+            with open(fl.abs_src_path, 'w', encoding='utf-8-sig') as f:
                 f.write(md_src)
             # Now read the file.
             pg.read_source(cfg)
@@ -625,7 +665,7 @@ class PageTests(unittest.TestCase):
         cfg = load_config()
         fl = File('missing.md', cfg['docs_dir'], cfg['site_dir'], cfg['use_directory_urls'])
         pg = Page('Foo', fl, cfg)
-        self.assertRaises(IOError, pg.read_source, cfg)
+        self.assertRaises(OSError, pg.read_source, cfg)
 
 
 class SourceDateEpochTests(unittest.TestCase):
@@ -647,7 +687,7 @@ class SourceDateEpochTests(unittest.TestCase):
             del os.environ['SOURCE_DATE_EPOCH']
 
 
-class RelativePathExtensionTests(LogTestCase):
+class RelativePathExtensionTests(unittest.TestCase):
 
     DOCS_DIR = os.path.join(os.path.abspath(os.path.dirname(__file__)), '../integration/subpages/docs')
 
@@ -661,112 +701,112 @@ class RelativePathExtensionTests(LogTestCase):
         pg.render(cfg, Files(fs))
         return pg.content
 
-    @mock.patch('io.open', mock.mock_open(read_data='[link](non-index.md)'))
+    @mock.patch('mkdocs.structure.pages.open', mock.mock_open(read_data='[link](non-index.md)'))
     def test_relative_html_link(self):
         self.assertEqual(
             self.get_rendered_result(['index.md', 'non-index.md']),
             '<p><a href="non-index/">link</a></p>'  # No trailing /
         )
 
-    @mock.patch('io.open', mock.mock_open(read_data='[link](index.md)'))
+    @mock.patch('mkdocs.structure.pages.open', mock.mock_open(read_data='[link](index.md)'))
     def test_relative_html_link_index(self):
         self.assertEqual(
             self.get_rendered_result(['non-index.md', 'index.md']),
             '<p><a href="..">link</a></p>'
         )
 
-    @mock.patch('io.open', mock.mock_open(read_data='[link](sub2/index.md)'))
+    @mock.patch('mkdocs.structure.pages.open', mock.mock_open(read_data='[link](sub2/index.md)'))
     def test_relative_html_link_sub_index(self):
         self.assertEqual(
             self.get_rendered_result(['index.md', 'sub2/index.md']),
             '<p><a href="sub2/">link</a></p>'  # No trailing /
         )
 
-    @mock.patch('io.open', mock.mock_open(read_data='[link](sub2/non-index.md)'))
+    @mock.patch('mkdocs.structure.pages.open', mock.mock_open(read_data='[link](sub2/non-index.md)'))
     def test_relative_html_link_sub_page(self):
         self.assertEqual(
             self.get_rendered_result(['index.md', 'sub2/non-index.md']),
             '<p><a href="sub2/non-index/">link</a></p>'  # No trailing /
         )
 
-    @mock.patch('io.open', mock.mock_open(read_data='[link](file%20name.md)'))
+    @mock.patch('mkdocs.structure.pages.open', mock.mock_open(read_data='[link](file%20name.md)'))
     def test_relative_html_link_with_encoded_space(self):
         self.assertEqual(
             self.get_rendered_result(['index.md', 'file name.md']),
             '<p><a href="file%20name/">link</a></p>'
         )
 
-    @mock.patch('io.open', mock.mock_open(read_data='[link](file name.md)'))
+    @mock.patch('mkdocs.structure.pages.open', mock.mock_open(read_data='[link](file name.md)'))
     def test_relative_html_link_with_unencoded_space(self):
         self.assertEqual(
             self.get_rendered_result(['index.md', 'file name.md']),
             '<p><a href="file%20name/">link</a></p>'
         )
 
-    @mock.patch('io.open', mock.mock_open(read_data='[link](../index.md)'))
+    @mock.patch('mkdocs.structure.pages.open', mock.mock_open(read_data='[link](../index.md)'))
     def test_relative_html_link_parent_index(self):
         self.assertEqual(
             self.get_rendered_result(['sub2/non-index.md', 'index.md']),
             '<p><a href="../..">link</a></p>'
         )
 
-    @mock.patch('io.open', mock.mock_open(read_data='[link](non-index.md#hash)'))
+    @mock.patch('mkdocs.structure.pages.open', mock.mock_open(read_data='[link](non-index.md#hash)'))
     def test_relative_html_link_hash(self):
         self.assertEqual(
             self.get_rendered_result(['index.md', 'non-index.md']),
             '<p><a href="non-index/#hash">link</a></p>'
         )
 
-    @mock.patch('io.open', mock.mock_open(read_data='[link](sub2/index.md#hash)'))
+    @mock.patch('mkdocs.structure.pages.open', mock.mock_open(read_data='[link](sub2/index.md#hash)'))
     def test_relative_html_link_sub_index_hash(self):
         self.assertEqual(
             self.get_rendered_result(['index.md', 'sub2/index.md']),
             '<p><a href="sub2/#hash">link</a></p>'
         )
 
-    @mock.patch('io.open', mock.mock_open(read_data='[link](sub2/non-index.md#hash)'))
+    @mock.patch('mkdocs.structure.pages.open', mock.mock_open(read_data='[link](sub2/non-index.md#hash)'))
     def test_relative_html_link_sub_page_hash(self):
         self.assertEqual(
             self.get_rendered_result(['index.md', 'sub2/non-index.md']),
             '<p><a href="sub2/non-index/#hash">link</a></p>'
         )
 
-    @mock.patch('io.open', mock.mock_open(read_data='[link](#hash)'))
+    @mock.patch('mkdocs.structure.pages.open', mock.mock_open(read_data='[link](#hash)'))
     def test_relative_html_link_hash_only(self):
         self.assertEqual(
             self.get_rendered_result(['index.md']),
             '<p><a href="#hash">link</a></p>'
         )
 
-    @mock.patch('io.open', mock.mock_open(read_data='![image](image.png)'))
+    @mock.patch('mkdocs.structure.pages.open', mock.mock_open(read_data='![image](image.png)'))
     def test_relative_image_link_from_homepage(self):
         self.assertEqual(
             self.get_rendered_result(['index.md', 'image.png']),
             '<p><img alt="image" src="image.png" /></p>'  # no opening ./
         )
 
-    @mock.patch('io.open', mock.mock_open(read_data='![image](../image.png)'))
+    @mock.patch('mkdocs.structure.pages.open', mock.mock_open(read_data='![image](../image.png)'))
     def test_relative_image_link_from_subpage(self):
         self.assertEqual(
             self.get_rendered_result(['sub2/non-index.md', 'image.png']),
             '<p><img alt="image" src="../../image.png" /></p>'
         )
 
-    @mock.patch('io.open', mock.mock_open(read_data='![image](image.png)'))
+    @mock.patch('mkdocs.structure.pages.open', mock.mock_open(read_data='![image](image.png)'))
     def test_relative_image_link_from_sibling(self):
         self.assertEqual(
             self.get_rendered_result(['non-index.md', 'image.png']),
             '<p><img alt="image" src="../image.png" /></p>'
         )
 
-    @mock.patch('io.open', mock.mock_open(read_data='*__not__ a link*.'))
+    @mock.patch('mkdocs.structure.pages.open', mock.mock_open(read_data='*__not__ a link*.'))
     def test_no_links(self):
         self.assertEqual(
             self.get_rendered_result(['index.md']),
             '<p><em><strong>not</strong> a link</em>.</p>'
         )
 
-    @mock.patch('io.open', mock.mock_open(read_data='[link](non-existant.md)'))
+    @mock.patch('mkdocs.structure.pages.open', mock.mock_open(read_data='[link](non-existant.md)'))
     def test_bad_relative_html_link(self):
         with self.assertLogs('mkdocs', level='WARNING') as cm:
             self.assertEqual(
@@ -779,21 +819,21 @@ class RelativePathExtensionTests(LogTestCase):
              "to 'non-existant.md' which is not found in the documentation files."]
         )
 
-    @mock.patch('io.open', mock.mock_open(read_data='[external link](http://example.com/index.md)'))
+    @mock.patch('mkdocs.structure.pages.open', mock.mock_open(read_data='[external](http://example.com/index.md)'))
     def test_external_link(self):
         self.assertEqual(
             self.get_rendered_result(['index.md']),
-            '<p><a href="http://example.com/index.md">external link</a></p>'
+            '<p><a href="http://example.com/index.md">external</a></p>'
         )
 
-    @mock.patch('io.open', mock.mock_open(read_data='[absolute link](/path/to/file.md)'))
+    @mock.patch('mkdocs.structure.pages.open', mock.mock_open(read_data='[absolute link](/path/to/file.md)'))
     def test_absolute_link(self):
         self.assertEqual(
             self.get_rendered_result(['index.md']),
             '<p><a href="/path/to/file.md">absolute link</a></p>'
         )
 
-    @mock.patch('io.open', mock.mock_open(read_data='<mail@example.com>'))
+    @mock.patch('mkdocs.structure.pages.open', mock.mock_open(read_data='<mail@example.com>'))
     def test_email_link(self):
         self.assertEqual(
             self.get_rendered_result(['index.md']),