+++ mkdocs/tests/theme_tests.py
@@ -1,9 +1,7 @@
-from __future__ import unicode_literals
-
 import os
 import tempfile
 import unittest
-import mock
+from unittest import mock
 
 import mkdocs
 from mkdocs.theme import Theme
@@ -16,7 +14,7 @@ theme_dir = os.path.abspath(os.path.join(mkdocs_dir, 'themes'))
 
 def get_vars(theme):
     """ Return dict of theme vars. """
-    return dict([(k, theme[k]) for k in iter(theme)])
+    return {k: theme[k] for k in iter(theme)}
 
 
 class ThemeTests(unittest.TestCase):
@@ -27,14 +25,15 @@ class ThemeTests(unittest.TestCase):
             theme.dirs,
             [os.path.join(theme_dir, 'mkdocs'), mkdocs_templates_dir]
         )
-        self.assertEqual(theme.static_templates, set(['404.html', 'sitemap.xml']))
+        self.assertEqual(theme.static_templates, {'404.html', 'sitemap.xml'})
         self.assertEqual(get_vars(theme), {
             'include_search_page': False,
             'search_index_only': False,
             'highlightjs': True,
             'hljs_style': 'github',
             'hljs_languages': [],
-            'shortcuts': {'help': 191, 'next': 78, 'previous': 80, 'search': 83}
+            'shortcuts': {'help': 191, 'next': 78, 'previous': 80, 'search': 83},
+            'nav_style': 'primary'
         })
 
     def test_custom_dir(self):
@@ -61,7 +60,7 @@ class ThemeTests(unittest.TestCase):
         theme = Theme(name='mkdocs', static_templates='foo.html')
         self.assertEqual(
             theme.static_templates,
-            set(['404.html', 'sitemap.xml', 'foo.html'])
+            {'404.html', 'sitemap.xml', 'foo.html'}
         )
 
     def test_vars(self):
@@ -78,7 +77,7 @@ class ThemeTests(unittest.TestCase):
     def test_no_theme_config(self, m):
         theme = Theme(name='mkdocs')
         self.assertEqual(m.call_count, 1)
-        self.assertEqual(theme.static_templates, set(['sitemap.xml']))
+        self.assertEqual(theme.static_templates, {'sitemap.xml'})
 
     def test_inherited_theme(self):
         m = mock.Mock(side_effect=[
@@ -97,5 +96,5 @@ class ThemeTests(unittest.TestCase):
                 ]
             )
             self.assertEqual(
-                theme.static_templates, set(['sitemap.xml', 'child.html', 'parent.html'])
+                theme.static_templates, {'sitemap.xml', 'child.html', 'parent.html'}
             )